+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.2] 2025-10-14
+### Changed
+- If Master's floruit dates are provided in place of birth/death dates, show
+those dates and years of activity.
+### Fixed
+- Master's successor listing has been restored.
+
 ## [1.7.1] 2025-10-13
 ### Fixed
 - Fix auto-save value for master properties like `uuid ` and `name_latin`
@@ -245,7 +252,8 @@ the slug's name
 - Initial `index.php`, `style.css`, and `functions.php`
 - This Changelog
 
-[Unreleased]: https://github.com/erremauro/cigno-zen/compare/v1.7.1...HEAD
+[Unreleased]: https://github.com/erremauro/cigno-zen/compare/v1.7.2...HEAD
+[1.7.2]: https://github.com/erremauro/cigno-zen/releases/tag/v1.7.2
 [1.7.1]: https://github.com/erremauro/cigno-zen/releases/tag/v1.7.1
 [1.7.0]: https://github.com/erremauro/cigno-zen/releases/tag/v1.7.0
 [1.6.2]: https://github.com/erremauro/cigno-zen/releases/tag/v1.6.2